requirements/requirements-dev.txt
fastapi>=0.110
uvicorn[standard]>=0.23
python-multipart>=0.0.6
pydantic>=2.11
pydantic-settings>=2.10
python-dotenv>=1.0
numpy>=2.3
pandas>=2.3
scikit-learn>=1.7
sentence-transformers>=5.0
faiss-cpu>=1.12
pdfplumber>=0.11
PyPDF2>=3.0
beautifulsoup4>=4.12
requests>=2.31
streamlit>=1.32
matplotlib>=3.7
pytest>=8.0
build>=1.0
coloredlogs>=15.0
rich>=14.0
watchdog>=6.0
